Fixing game config download, save and load
Download from web happens only on app launch and resume.
Game config info is only loaded from file. This will happen immediately
and not wait for download from web to finish. It is not necessary to
have completely up-to-date config info every time.
If file doesn't exist we will load from built-in gamestate.txt (this
will occur on first use).