Bugs from No asset and No wifi playthroughs
1. Don't close Out of Puzzle user control on overlay click.
2. BoostsUserControl.cs: set a default AnswerPuzzleCost if the current
puzzle can't be loaded. This value will never be used.
3. Fix LoadCurrentPuzzle logic to ensure at least either OutOfPuzzles or
NewPuzzles dialog show up if the current puzzle can't be loaded.
4. Changes to the DownloadManager to download the CurrentPuzzle pictures
over other pictures in the queue.
5. NewPuzzlesUserControl.cs: Only reload the game when all files in the
DownloadManager have finished downloading. This is fine as if we don't
have all the files for the CurrentPuzzle we change the DownloadManager
queue to be only the files of the CurrentPuzzle.